requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
cffi==1.15.1
channels==4.0.0
cryptography==37.0.4
cycler==0.11.0
dj-database-url==1.0.0
Django==4.1
django-crispy-forms==1.14.0
django-filter==22.1
django-redis==5.2.0
django-rest-knox==4.2.0
djangochannelsrestframework==1.1.0
djangorestframework==3.13.1
fonttools==4.37.1
gunicorn==20.1.0
Jinja2==3.1.2
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
mysqlclient==2.1.1
numpy==1.23.3
packaging==21.3
pandas==1.4.4
Pillow==9.2.0
psycopg2-binary==2.9.3
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
redis==4.5.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
